Change longs to long doubles to allow decimal notation

+++ src/headers/eval.h
@@ -3,6 +3,6 @@
 
 #include "opstack.h"
 
-extern long evaluateOpStack(OpStack *input);
+extern long double evaluateOpStack(OpStack *input);
 
 #endif
